# contracts/openapi.yaml
openapi: 3.1.0
info:
  title: MTB Galaxy API
  version: 1.0.0
paths:
...
  /v1/users:
    post:
      tags:
      - users
      summary: Create User
      operationId: create_user_v1_users_post
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserByNicknameRequest'
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserByNicknameResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
    get:
      tags:
      - users
      summary: Get User By Nickname
      operationId: get_user_by_nickname_v1_users_get
      parameters:
      - name: nickname
        in: query
        required: true
        schema:
          type: string
          title: Nickname
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserByNicknameResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /v1/me:
    get:
      tags:
      - users
      summary: Get Me
      operationId: get_me_v1_me_get
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MeResponse'
      security:
      - HTTPBearer: []
...
  /v1/planets/list:
    get:
      tags:
      - planets
      summary: Get Planets
      operationId: get_planets_v1_planets_list_get
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                items:
                  $ref: '#/components/schemas/PlanetListItemOut'
                type: array
                title: Response Get Planets V1 Planets List Get
      security:
      - HTTPBearer: []
  /v1/planets/{planet_id}/progress:
    get:
      tags:
      - planets
      summary: Get Planet Progress
      operationId: get_planet_progress_v1_planets__planet_id__progress_get
      security:
      - HTTPBearer: []
      parameters:
      - name: planet_id
        in: path
        required: true
        schema:
          type: string
          title: Planet Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/modules__planets__schemas__PlanetProgressOut'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /v1/planets/{planet_id}/focus:
    patch:
      tags:
      - planets
      summary: Focus Planet
      operationId: focus_planet_v1_planets__planet_id__focus_patch
      security:
      - HTTPBearer: []
      parameters:
      - name: planet_id
        in: path
        required: true
        schema:
          type: string
          title: Planet Id
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/modules__planets__schemas__FocusPlanetRequest'
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FocusPlanetResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
...
components:
  schemas:
...
    FocusPlanetResponse:
      properties:
        planet_id:
          type: string
          title: Planet Id
        focus:
          type: boolean
          title: Focus
      type: object
      required:
      - planet_id
      - focus
      title: FocusPlanetResponse
...
    HTTPValidationError:
      properties:
        detail:
          items:
            $ref: '#/components/schemas/ValidationError'
          type: array
          title: Detail
      type: object
      title: HTTPValidationError
...
    MeResponse:
      properties:
        id:
          type: string
          title: Id
        phone:
          type: string
          title: Phone
        name:
          type: string
          title: Name
        nickname:
          anyOf:
          - type: string
          - type: 'null'
          title: Nickname
        daily_game_attempts_used:
          type: integer
          title: Daily Game Attempts Used
        daily_game_attempts_limit:
          type: integer
          title: Daily Game Attempts Limit
        user:
          $ref: '#/components/schemas/UserSummaryOut'
        selected_planet:
          $ref: '#/components/schemas/PlanetCode'
      type: object
      required:
      - id
      - phone
      - name
      - daily_game_attempts_used
      - daily_game_attempts_limit
      - user
      - selected_planet
      title: MeResponse
...
    PlanetCode:
      type: string
      enum:
      - ORBIT_COMMERCE
      - CREDIT_SHIELD
      - SOCIAL_RING
      title: PlanetCode
    PlanetGameOut:
      properties:
        code:
          type: string
          title: Code
        name:
          type: string
          title: Name
        attempts_used:
          type: integer
          title: Attempts Used
        attempts_limit:
          type: integer
          title: Attempts Limit
      type: object
      required:
      - code
      - name
      - attempts_used
      - attempts_limit
      title: PlanetGameOut
...
    PlanetListItemOut:
      properties:
        id:
          type: string
          title: Id
        name:
          type: string
          title: Name
        cashback_percent:
          type: number
          title: Cashback Percent
        progress_percent:
          type: number
          title: Progress Percent
      type: object
      required:
      - id
      - name
      - cashback_percent
      - progress_percent
      title: PlanetListItemOut
...
    SegmentKey:
      type: string
      enum:
      - student
      - first-jobber
      - freelancer
      title: SegmentKey
...
    UserByNicknameRequest:
      properties:
        nickname:
          type: string
          title: Nickname
        user_id:
          anyOf:
          - type: string
          - type: 'null'
          title: User Id
      type: object
      required:
      - nickname
      title: UserByNicknameRequest
    UserByNicknameResponse:
      properties:
        id:
          type: string
          title: Id
        nickname:
          type: string
          title: Nickname
      type: object
      required:
      - id
      - nickname
      title: UserByNicknameResponse
    UserSummaryOut:
      properties:
        user_id:
          type: string
          title: User Id
        phone:
          type: string
          title: Phone
        display_name:
          type: string
          title: Display Name
        nickname:
          anyOf:
          - type: string
          - type: 'null'
          title: Nickname
        segment:
          $ref: '#/components/schemas/SegmentKey'
        created_at:
          type: string
          format: date-time
          title: Created At
        id:
          type: string
          title: Id
          readOnly: true
        name:
          type: string
          title: Name
          readOnly: true
      type: object
      required:
      - user_id
      - phone
      - display_name
      - segment
      - created_at
      - id
      - name
      title: UserSummaryOut
    ValidationError:
      properties:
        loc:
          items:
            anyOf:
            - type: string
            - type: integer
          type: array
          title: Location
        msg:
          type: string
          title: Message
        type:
          type: string
          title: Error Type
        input:
          title: Input
        ctx:
          type: object
          title: Context
      type: object
      required:
      - loc
      - msg
      - type
      title: ValidationError
...
    modules__planets__schemas__ConstellationOut:
      properties:
        name:
          type: string
          title: Name
        index:
          type: integer
          title: Index
        big_stars:
          type: integer
          title: Big Stars
        current_big_star:
          type: integer
          title: Current Big Star
        small_stars_current:
          type: integer
          title: Small Stars Current
        small_stars_required:
          type: integer
          title: Small Stars Required
        big_stars_state:
          items:
            type: boolean
          type: array
          title: Big Stars State
        small_stars_state:
          items:
            type: boolean
          type: array
          title: Small Stars State
      type: object
      required:
      - name
      - index
      - big_stars
      - current_big_star
      - small_stars_current
      - small_stars_required
      - big_stars_state
      - small_stars_state
      title: ConstellationOut
    modules__planets__schemas__FocusPlanetRequest:
      properties:
        focus:
          type: boolean
          title: Focus
      type: object
      required:
      - focus
      title: FocusPlanetRequest
    modules__planets__schemas__PlanetProgressOut:
      properties:
        planet_id:
          type: string
          title: Planet Id
        name:
          type: string
          title: Name
        cashback_percent:
          type: number
          title: Cashback Percent
        max_cashback_reached:
          type: boolean
          title: Max Cashback Reached
        constellation:
          $ref: '#/components/schemas/modules__planets__schemas__ConstellationOut'
        period_small_stars:
          type: integer
          title: Period Small Stars
        big_stars_until_increase:
          type: integer
          title: Big Stars Until Increase
        game:
          $ref: '#/components/schemas/PlanetGameOut'
      type: object
      required:
      - planet_id
      - name
      - cashback_percent
      - max_cashback_reached
      - constellation
      - period_small_stars
      - big_stars_until_increase
      - game
      title: PlanetProgressOut
...
  securitySchemes:
    HTTPBearer:
      type: http
      scheme: bearer